+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # REPLACE-GLOBAL-WITH-NEXT-VERSION
+- MIG-125 - Improves the migration of orders, that customer order count is indexed
 - MIG-181 - Provide migration of main variant information for Shopware 5.4 / 5.6
 
 # 2.2.2